+++ pyjt/Errors.py
@@ -1,3 +1,4 @@
 
 class ElementNotFoundError(Exception):
+    """ No element was found matching the given search criteria. """
     pass