+++ pyjt/Fixture.py
@@ -20,13 +20,18 @@ class Fixture:
         """ Search a control as a sub-element of the control
             managed by this fixture.
 
-            Parameters
-            ----------
+            Args:
+                locator (Locator):  A locator object to find the sub-control
+                **kwargs:           Additional search parameters.
 
-            locator:    Locator
-                A locator object to find the sub-control
-            **kwargs:
-                Additional search parameters.
+            Returns:
+                Fixture:
+                    A fixture pointing to the first component
+                    matching the search criteria.
+
+            Raises:
+                ElementNotFoundError:
+                    No element was found matching the search criteria
         """
         log.debug(f"frame.find({locator}, {kwargs})")
         locator = locator if locator else Locator(**kwargs)
@@ -36,6 +41,24 @@ class Fixture:
         return Fixture(control)
 
     def find(self, role, **kwargs):
+        """ Search a control of type **role** as a sub-control of this control.
+
+            Args:
+                role (Class):
+                    The class type of the component to look for, e.g.
+                    javax.swing.JTextField
+                **kwargs:
+                    Search criteria for the component.
+
+            Returns:
+                Fixture:
+                    A fixture pointing to the first component
+                    matching the search criteria.
+
+            Raises:
+                ElementNotFoundError:
+                    No element was found matching the search criteria
+        """
         kwargs['role'] = role
         return self.locate(**kwargs)
 
@@ -76,6 +99,13 @@ class Fixture:
             self.robot.type(text)
         if mode == FillMode.SET:
             self._control.setText(text)
+
+    @property
+    def control(self):
+        """ Returns:
+                Proxy:  A reference to the control managed by this fixture.
+        """
+        return self._control
 
     def __getattr__(self, name):
         return getattr(self._control, name)