+++ RELEASES.md
@@ -1,3 +1,11 @@
+# v0.4.0
+> 2023-08-04
+
+-   Add `Proxy.isinstance()` function
+-   Rename `Proxy.instance` property to `Proxy.object`
+-   Add `Fixture.object` property
+-   Improve docmentation.
+
 # v0.3.1
 > 2023-08-03